fix: bug with folder

+++ 04-copy-directory/index.js
@@ -11,15 +11,19 @@ fs.access(filesCopy, fs.F_OK, (err) => {
       fs.readdir(filesPath, 'utf8', (err, files) => {
         if (err) throw err;
         files.forEach((file)=>{
-          fs.readFile(path.join(filesPath, file), 'utf-8', (err, data) => {
-            if(err){
-              console.log('err');
+          fs.stat(path.join(filesPath, file),(err, data) => {
+            if(data.isFile()){
+              fs.readFile(path.join(filesPath, file), 'utf-8', (err, data) => {
+                if(err){
+                  console.log('err');
+                }
+                fs.writeFile(path.join(filesCopy, file), data, (err) => {
+                  if (err) {
+                    throw err;
+                  } });
+              });
             }
-            fs.writeFile(path.join(filesCopy, file), data, (err) => {
-              if (err) {
-                throw err;
-              } });
-          });
+          });      
         });
       });
     });
@@ -29,15 +33,19 @@ fs.access(filesCopy, fs.F_OK, (err) => {
     fs.readdir(filesPath, 'utf8', (err, files) => {
       if (err) throw err;
       files.forEach((file)=>{
-        fs.readFile(path.join(filesPath, file), 'utf-8', (err, data) => {
-          if(err){
-            console.log('err');
+        fs.stat(path.join(filesPath, file),(err, data) => {
+          if(data.isFile()){
+            fs.readFile(path.join(filesPath, file), 'utf-8', (err, data) => {
+              if(err){
+                console.log('err');
+              }
+              fs.writeFile(path.join(filesCopy, file), data, (err) => {
+                if (err) {
+                  throw err;
+                } });
+            });
           }
-          fs.writeFile(path.join(filesCopy, file), data, (err) => {
-            if (err) {
-              throw err;
-            } });
-        });
+        }); 
       });
     });
   }
